% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{c_CV}
\alias{c_CV}
\title{c_CV: Coefficient of Variation (CV) of false discovery rate (c).}
\usage{
c_CV(n, c)
}
\arguments{
\item{n}{\itemize{
\item Number of trials used to estimate c
}}

\item{c}{\itemize{
\item false discovery rate (proportion of false positive detections
divided by the total number predicted positive)
}}
}
\value{
cv.c - coefficient of variation of the false discovery rate
}
\description{
CV.c is calculated from the variance of c, which depends on the probability
of false alarm (c), a proportion.
}
\details{
Rationale for this calculation Variance of a proportion, c derived from a
binomial process. Zar 2010 Eqn. 24.16 indicates sigma^2 = pq/n
here p = c, q = (1-c), and n is the number of detections inspected when
estimating c, so:
\enumerate{
\item var(c)= c * (1-c)/n
}

And required inputs are n, the number of trials conducted (i.e. how many
detections were inspected when estimating c) and the estimate of c (as a
proportion between 0 and 1 inclusive).
}
