% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensity-package.R
\docType{package}
\name{callDensity-package}
\alias{callDensity}
\alias{callDensity-package}
\title{callDensity: Analysis for bioacoustic call density estimation (ABCDE)}
\description{
Software for estimating the density of animal vocalisations (calls) This package includes functions for estimating call density of animal sounds, especially those of low-frequency baleen whales, by applying the algorithms described by Danielle Harris. In particular these algorithms can be used to estimate call density using auxiliary information to estimate the probability of detection within an area. The specific auxiliary information required by this package comes from the sonar equation, specifically, distributions of source levels (SL), transmission losses (TL), and noise levels (NL).
}
\keyword{internal}
