% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{capHistTimeSeason}
\alias{capHistTimeSeason}
\title{Add R datetime, season, and month codes to a capture history table}
\usage{
capHistTimeSeason(x)
}
\arguments{
\item{x}{Capture history table with column t0_tableN of Matlab datenum's}
}
\value{
Capture history table with column t, season, and month containing
POSIXct, and timeCode factors for season and month.
}
\description{
Add R datetime, season, and month codes to a capture history table
}
