% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{capHistTimeSeason2}
\alias{capHistTimeSeason2}
\title{Add R datetime and season to a capture history table (version 2)}
\usage{
capHistTimeSeason2(x)
}
\arguments{
\item{x}{A capture history table with column t0 containing a matlab datenum}
}
\value{
A capture history table with original columns plus:
    t: a posixCT corresponding to the date and time of t0
    season: a factor containing summer, autumn, winter, spring, or year
      corresponding to the season of t0
    month: a factor containing '01'-'12' corresponding to the month of t0
}
\description{
Add R datetime and season to a capture history table (version 2)
}
