% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{ciFromCV}
\alias{ciFromCV}
\title{Confidence interval from density and it's CV}
\usage{
ciFromCV(a, cv)
}
\arguments{
\item{a}{abundance (or density)}

\item{cv}{coefficient of variation, in decimal format, i.e., a CV of 10%
would be 0.1.}
}
\value{
list with lower and upper 95% confidence intervals
}
\description{
Quick function to convert an abundance with a coefficient of variation (CV)
into a 95% confidence interval.
}
