% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{collectDensityResults}
\alias{collectDensityResults}
\title{Load specified density results text files in a directory}
\usage{
collectDensityResults(path, densityString = paste0("^density_", siteCode))
}
\arguments{
\item{path}{Location (path) of text files containing call density results}

\item{densityString}{Prefix of _density files to be loaded}
}
\value{
Data.frame of all call density results in the specified path
}
\description{
Load specified density results text files in a directory
}
