% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{Dc_CV}
\alias{Dc_CV}
\title{Coefficient of Variation (CV) for call density (Dc)}
\usage{
Dc_CV(CV.Nc, CV.pa, CV.c)
}
\arguments{
\item{CV.Nc}{- CV of number of calls Nc}

\item{CV.pa}{- CV of probability of detection in the area, pa}

\item{CV.c}{- CV of the false discovery rate, c}
}
\value{
CV.Dc - coefficient of variation of the call density
}
\description{
The uncertainty (CV) in the estimated call density (Dc) is calculated using
the Delta Method, which combines the CV of the individual parameters (Nc, c
and pa).
}
\details{
Here, the uncertainties of Nc, c, and pa are assumed to be independent, given
the fundamental premise of the Delta Method in this application.
}
