% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{defaultOutputFileNames}
\alias{defaultOutputFileNames}
\title{Default names of output files for call density estimation}
\usage{
defaultOutputFileNames(p, season)
}
\arguments{
\item{p}{Data.frame containing parameters for call density estimation}

\item{season}{TimeCode to filter month, season, or full year}
}
\value{
Parameter file with file names corresponding to the correct timeCode
}
\description{
Given data.frame of call density parameters, p, create default parameters and
output file-names for Monte-Carlo modelling and estimating call-density.
Append all of these parameters to the data.frame and return it.
}
