% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{density.results.table}
\alias{density.results.table}
\title{Table of call densities including CVs (ggplot2)}
\usage{
\method{density}{results.table}(d)
}
\arguments{
\item{d}{Data.frame of call density results#'}
}
\description{
Table of call densities including CVs (ggplot2)
}
