% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{densityPlot}
\alias{densityPlot}
\title{Bar-plot of call density by timeCodes (ggplot2)}
\usage{
densityPlot(d)
}
\arguments{
\item{d}{Data.frame of call density results}
}
\description{
Bar-plot of call density by timeCodes (ggplot2)
}
