% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{densityResultsTable}
\alias{densityResultsTable}
\title{Table of call densities including CVs (ggplot2)}
\usage{
densityResultsTable(d)
}
\arguments{
\item{d}{Data.frame of call density results#'}
}
\description{
Table of call densities including CVs (ggplot2)
}
