% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{deploymentDuration}
\alias{deploymentDuration}
\title{Calculate deployment duration from recording start times and durations}
\usage{
deploymentDuration(duration, startDate, season = "year")
}
\arguments{
\item{duration}{- list of recording durations (in s)}

\item{startDate}{- list of start dates for each duration}

\item{season}{- Subset the startDate by this season/month/year (default='year')}
}
\value{
- duration of effort (as numeric)
}
\description{
Calculate deployment duration from recording start times and durations
}
