% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{deploymentDurationFromsoundFolderCsv}
\alias{deploymentDurationFromsoundFolderCsv}
\title{Calculate deployment duration from a Matlab soundFolder csv file.}
\usage{
deploymentDurationFromsoundFolderCsv(fullYearEffortFile, season = "year")
}
\arguments{
\item{fullYearEffortFile}{\itemize{
\item csv file containing a row for each time period
that was included in the study and columns named startDate and duration.
startDate should be a matlab datenum, and duration should be numeric
and in seconds.
}}

\item{season}{\itemize{
\item default='year'
}}
}
\description{
Calculate deployment duration from a Matlab soundFolder csv file.
}
