% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{detection.rate.plot}
\alias{detection.rate.plot}
\title{Bar-plot of detection rates by timeCodes (ggplot2)}
\usage{
detection.rate.plot(d)
}
\arguments{
\item{d}{Data.frame of call density results}
}
\description{
Bar-plot of detection rates by timeCodes (ggplot2)
}
