% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{detectionRateCorrectedPlot}
\alias{detectionRateCorrectedPlot}
\title{Bar-plot of detection rates by timeCodes and scaled (corrected) by precision (ggplot2)}
\usage{
detectionRateCorrectedPlot(d)
}
\arguments{
\item{d}{Data.frame of call density results}
}
\description{
Bar-plot of detection rates by timeCodes and scaled (corrected) by precision (ggplot2)
}
