% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{detectionRatePlot}
\alias{detectionRatePlot}
\title{Bar-plot of detection rates by timeCodes (ggplot2)}
\usage{
detectionRatePlot(d)
}
\arguments{
\item{d}{Data.frame of call density results}
}
\description{
Bar-plot of detection rates by timeCodes (ggplot2)
}
