% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDetFun.R
\name{detMissSNR}
\alias{detMissSNR}
\title{Split an SNRinfo data.frame into detected/missed SNR vectors}
\usage{
detMissSNR(obs)
}
\arguments{
\item{obs}{Data.frame with columns \code{SNR} and \code{Detected}.}
}
\value{
A list with elements \code{det} and \code{miss}.
}
\description{
Split an SNRinfo data.frame into detected/missed SNR vectors
}
\keyword{internal}
