% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{falseDiscoveryCV}
\alias{falseDiscoveryCV}
\title{CV of false discovery rate using Cochran approximation}
\usage{
falseDiscoveryCV(p, season)
}
\arguments{
\item{season}{}
}
\value{
CV of false discovery rate using Cochran approximation
}
\description{
CV of false discovery rate using Cochran approximation
}
