% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{falseDiscoveryRateFromNth}
\alias{falseDiscoveryRateFromNth}
\title{False discovery rate from inspection of every Nth detection}
\usage{
falseDiscoveryRateFromNth(falsePositiveXlsx, season = "year")
}
\arguments{
\item{falsePositiveXlsx}{Excel spreadsheet with timestamp of false positives.
This spreadsheet requires two columns: one called UTC and the other called
'True Positive Rate.' These should be in a worksheet called 'conference'.}

\item{season}{Month or season over which to subset the data. Months can be
01-12, and seasons can be 'summer','autumn','winter','spring', or 'year'.}
}
\description{
False discovery rate from inspection of every Nth detection
}
