% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDetFun.R
\name{fitDetFun}
\alias{fitDetFun}
\title{Fit an SNR detection function}
\usage{
fitDetFun(
  SNRinfo,
  modelType = c("gam", "glm", "scam", "vglm"),
  numKnots = 3,
  yColNames = c("detect_observer1", "detect_observer2"),
  whichObserver = NULL
)
}
\arguments{
\item{SNRinfo}{data.frame containing SNR and detection information.}

\item{modelType}{character.}

\item{numKnots}{spline basis dimension for GAM/SCAM.}

\item{yColNames}{observer columns for capture-recapture models.}

\item{whichObserver}{observer used for prediction.}
}
\description{
Fits a detection function relating probability of detection to SNR.

Supported model types are:
\itemize{
\item glm
\item gam
\item scam
\item vglm (capture-recapture)
}

Fits a detection function relating probability of detection to SNR.

Supported model types are:
\itemize{
\item glm
\item gam
\item scam
\item vglm (capture-recapture)
}
}
