% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{fitSNRbySeason}
\alias{fitSNRbySeason}
\title{Model SNR-detection function with season as a factor (deprecated?)}
\usage{
fitSNRbySeason(SNRinfo, season = year, useGLM = TRUE, numKnots = 3)
}
\arguments{
\item{SNRinfo}{Data.frame containing columns SNR, Detected, and season}

\item{season}{Factor containing a timeCode}

\item{useGLM}{Boolean, if true fit a GLM instead of GAM}

\item{numKnots}{Number of knots to use in the GAM}
}
\description{
Model SNR-detection function with season as a factor (deprecated?)
}
