% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{fitSNRdetectionFunc}
\alias{fitSNRdetectionFunc}
\title{Fit an SNR-detection function}
\usage{
fitSNRdetectionFunc(SNRinfo, useGLM = TRUE, useSCAM = TRUE, numKnots = 3)
}
\arguments{
\item{SNRinfo}{A data.frame containing detection information in columns
Detected and SNR}

\item{useGLM}{Boolean - if true a GLM will be fit instead of a GAM}

\item{useSCAM}{Boolean - if true a shape constrained model will be fit instead of a GAM}

\item{numKnots}{Number of knots in the GAM or SCAM}
}
\description{
SNR detection functions are binomial models of the form detected ~ SNR. These
can be a GAM (default), GLM, or SCAM.
}
