% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{listTLFiles}
\alias{listTLFiles}
\title{Return a list of existing TL files associated with the call density}
\usage{
listTLFiles(p, season = "")
}
\arguments{
\item{p}{Data.frame containing parameters for call density estimation}

\item{season}{TimeCode to filter month, season, or full year}
}
\value{
a list of all transmission loss files
}
\description{
Return a list of existing TL files associated with the call density
}
