% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{mat2Rdate}
\alias{mat2Rdate}
\title{Convert matlab datenum to R POSIXct}
\usage{
mat2Rdate(x)
}
\arguments{
\item{x}{List of Matlab datenums}
}
\value{
posixCT corresponding to the Matlab datenum that was input
}
\description{
Convert matlab datenum to R POSIXct
}
