% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{Nc_CV}
\alias{Nc_CV}
\title{Nc_CV: Coefficient of Variation (CV) of the number of detected calls (Nc).}
\usage{
Nc_CV(Nc, pa)
}
\arguments{
\item{Nc}{- Number of calls detected in the dataset (see function
countDetections())}

\item{pa}{- Average probability of detection in the study area (see function
pDetInArea())}
}
\value{
cv.Nc - coefficient of variation of the number of detected calls
}
\description{
CV.Nc is calculated from the variance of Nc, which depends on the probability
of detection (pa). Here, Nc represents the number of detected calls, and
independence between detections is assumed.
}
