% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{nlFromSnrInfo}
\alias{nlFromSnrInfo}
\title{Estimate noise level distribution for MC simulation from the noise
measurements included in the snrInfo file}
\usage{
nlFromSnrInfo(snrInfo)
}
\arguments{
\item{SNRinfo}{Data.frame of snrInfo (e.g. like that generated by
capHist2snrInfo)}
}
\value{
Data.frame with 1 row and 3 columns containing parameterised
  distribution of noise levels. Column names are mean, sd, and sampleSize.
  Distribution assumed to be normal.
}
\description{
Estimate noise level distribution for MC simulation from the noise
measurements included in the snrInfo file
}
