% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{noiseLevelDistribution}
\alias{noiseLevelDistribution}
\title{Mean and standard deviation of noise levels by month, season, or year.}
\usage{
noiseLevelDistribution(nlFile, season = "year")
}
\arguments{
\item{nlFile}{File name of csv containing noise level times and values}

\item{season}{TimeCode for season, month, or 'year'}
}
\value{
The mean and standard deviation from the table of noise levels.
}
\description{
Read a table of noise levels containing a columns t and NL for time and
level, respectively.
}
