% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySim.R
\name{plotDetectionDistribution}
\alias{plotDetectionDistribution}
\title{SNR histogram of true & false positives, and false negatives for callDensity simulation}
\usage{
plotDetectionDistribution(sim)
}
\arguments{
\item{sim}{\itemize{
\item callDensity simulation data.frame
}}
}
\value{
\itemize{
\item ggplot object of type geom_histogram
}
}
\description{
SNR histogram of true & false positives, and false negatives for callDensity simulation
}
