% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySim.R
\name{plotSpatialDetections}
\alias{plotSpatialDetections}
\title{Plot the spatial detection density for a callDensity simulation}
\usage{
plotSpatialDetections(sim)
}
\arguments{
\item{sim}{\itemize{
\item callDensity simulation
}}
}
\value{
ggplot object of type geom_bin_2d
}
\description{
Plot the spatial detection density for a callDensity simulation
}
