% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{predVglmPDet}
\alias{predVglmPDet}
\title{Title}
\usage{
predVglmPDet(
  snrDetFun.vglm,
  snrs,
  whichObserver = snrDetFun.vglm@extra$whichObserver
)
}
\arguments{
\item{whichObserver}{}
}
\description{
Title
}
