% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{readCapHist}
\alias{readCapHist}
\title{Read a capture history csv file (e.g. created in Matlab)}
\usage{
readCapHist(capHistFile)
}
\arguments{
\item{capHistFile}{\itemize{
\item name of the csv or txt file that contains capture
histories. TODO: Document the required columns/file format better
}}
}
\value{
capture history table as data.frame
}
\description{
Read a capture history csv file (e.g. created in Matlab)
}
