% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{show.density.input.table}
\alias{show.density.input.table}
\title{Show table of inputs into call density estimate}
\usage{
show.density.input.table(d, separateCommon = TRUE)
}
\arguments{
\item{d}{Data.frame containing call density results}

\item{separateCommon}{Boolean - if true input parameters k, w, SLmean, SLsd
will be put into a separate table with only a single row.}
}
\description{
Show table of inputs into call density estimate
}
