% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySim.R
\name{simCallAcoustics}
\alias{simCallAcoustics}
\title{Simulate acoustic properties of calls}
\usage{
simCallAcoustics(
  sim,
  SL = data.frame(mean = 190, sd = 4, sampleSize = 350),
  NL = data.frame(mean = 84, sd = 4, sampleSize = n),
  TL = function(r) {
     20 * log(r)
 }
)
}
\arguments{
\item{sim}{\itemize{
\item data.frame containing simulated call locations (from
simCallLocations). Each location must have a distance.
}}

\item{SL}{\itemize{
\item Source level distribution. SL is a data.frame with columns named
'mean' and 'sd' that describe the mean and standard deviation of the
distribution (normal in dB) from which source levels will be generated for
each simulated call.
}}

\item{NL}{\itemize{
\item Noise level distribution. NL is a data.frame with columns named
'mean' and 'sd' that describe the mean and standard deviation of the
distribution (normal in dB) of noise levels from which noise levels will be
generated for each simulated call
}}

\item{TL}{\itemize{
\item a function that takes a single argument, vector r, and returns
transmission losses for the ranges in that vector. The default function is
spherical spreading: Default = function(r){20*log(r)}
}}
}
\value{
Simulation data.frame including columns that contain simulated
source levels (SL) noise levels (noiseRMSdB), and transmission losses (TL).
}
\description{
Simulate acoustic properties of calls
}
