% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySim.R
\name{simTLradials_20logR}
\alias{simTLradials_20logR}
\title{Simulate transmission loss (TL) radials following geometric (spherical)
spreading law.}
\usage{
simTLradials_20logR(maxRange, rangeStep, numTransects)
}
\arguments{
\item{maxRange}{- Length of each radial transect}

\item{rangeStep}{- increment between points on each transect where TL}

\item{numTransects}{- Number of radial transects (i.e. columns in result)}
}
\value{
TL -
}
\description{
Create a TL data.frame with a column for range and subsequent columns for
transmission loss, TL for radial transects around a recorder. Transects will
have a total length of maxRange with increments between points along the
transect of rangeStep. The number of radial transects is specified as
numTransects such that the angular resolution of the transects will be
360/numTransects.
}
