% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDetFun.R
\name{subsampleRug}
\alias{subsampleRug}
\title{Subsample a vector of rug points for large datasets}
\usage{
subsampleRug(x, n)
}
\description{
Subsample a vector of rug points for large datasets
}
\keyword{internal}
