% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySim.R
\name{subsampleSimInTime}
\alias{subsampleSimInTime}
\title{Subsample from a simulation at evenly spaced time intervals.}
\usage{
subsampleSimInTime(
  sim,
  minDate = min(sim$datetime),
  maxDate = max(sim$datetime),
  interval = "41 hour",
  duration = 3600
)
}
\arguments{
\item{sim}{\itemize{
\item simulation data.frame containing column a time column named
'datetime'
}}

\item{minDate}{\itemize{
\item posixCT indicating the start of the first subsample
}}

\item{maxDate}{\itemize{
\item posixCT indicating the start of the last subsample
}}

\item{interval}{\itemize{
\item difftime interval between subsamples (e.g. '41 hour')
}}

\item{duration}{\itemize{
\item numeric indicating the duration (in s) of each subsample
}}
}
\value{
subsampledSim - a simulation dataframe containing only rows from the
input that fall within the subsample
}
\description{
Subsample from a simulation at evenly spaced time intervals.
}
