% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{subsetByTimeCode}
\alias{subsetByTimeCode}
\title{Subset a data.frame by by months or season}
\usage{
subsetByTimeCode(df, dt, timeCode)
}
\arguments{
\item{df}{A data.frame}

\item{dt}{A posixCT datetime}

\item{timeCode}{A timecode indicating the season, month, or 'year' for all data}
}
\description{
Subset a data.frame by timeCode, where timeCode can be either a 'season'
(summer, autumn, winter, spring), a month characters ('01'-'12'), or 'year',
with 'year' the same as including all data
}
