% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{time2monthCode}
\alias{time2monthCode}
\title{Lookup month codes from a POSIXct}
\usage{
time2monthCode(x)
}
\arguments{
\item{x}{POSIXct}
}
\value{
Factor containing month codes '01'-'12' for months from Jan-Dec.
}
\description{
Lookup the month codes for a given POSIXct, x. MonthCodes are a factor from
'01'-'12' respectively for months from Jan-Dec
}
