% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{time2season}
\alias{time2season}
\title{Lookup the season for a given POSIXct, x.}
\usage{
time2season(x)
}
\arguments{
\item{x}{POSIXct}
}
\value{
Factor containing seasons summer, autumn, winter, or spring
}
\description{
Seasons are defined by 3 month
periods from Summer=(Dec,Jan,Feb); Autumn=(Mar,Apr,May), Winter=(Jun,Jul,Aug),
Spring=(Sep,Oct,Nov)
}
