% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{updateDetectionFolder}
\alias{updateDetectionFolder}
\title{Change path to detection csv files}
\usage{
updateDetectionFolder(p, newFolder)
}
\arguments{
\item{p}{Call density detector parameter data.frame}

\item{newFolder}{relative or absolute path to where the detection files are
located}
}
\value{
new detection parameters with filenames that have updated folders
}
\description{
Change path to detection csv files
}
