% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{var.wtd.mean.cochran}
\alias{var.wtd.mean.cochran}
\title{Variance of a weighted mean following Chochran 1977 definition}
\usage{
var.wtd.mean.cochran(x, w)
}
\arguments{
\item{x}{list of data}

\item{w}{weights for data}
}
\description{
Variance of a weighted mean following Chochran 1977 definition
}
