% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{c_CV}
\alias{c_CV}
\title{c_CV: Coefficient of Variation (CV) of false discovery rate (c).}
\usage{
c_CV(n, c)
}
\arguments{
\item{n}{\itemize{
\item Number of trials used to estimate c
}}

\item{c}{\itemize{
\item false discovery rate (proportion of false positive detections
divided by the total number predicted positive)
}}
}
\value{
cv.c - coefficient of variation of the false discovery rate
}
\description{
CV.c is calculated from the variance of c, which depends on the probability
of false alarm (c), a binomial process.
}
\details{
Rationale for this calculation Variance of a binomial process, pa, is: 1)
sigma^2= n * c * (1-c)

We require n, the number of trials conducted (i.e. how many detections were
inspected when estimating c) and the estimate of c
}
