% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{capHist2snrInfo}
\alias{capHist2snrInfo}
\title{Convert a capture history FILE of multiple detections into an SNR.Info file}
\usage{
capHist2snrInfo(snr, season = "year")
}
\arguments{
\item{season}{A timeCode corresponding to months, seasons, or year}

\item{capHistFile}{A capture history file containing detections and SNR info}
}
\description{
Convert a capture history FILE of multiple detections into an SNR.Info file
}
