% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{capHistTosnrInfo}
\alias{capHistTosnrInfo}
\title{Convert capture history DATA.FRAME into the 'SNRinfo' format used
by the callDensity package
capHistTosnrInfo}
\usage{
capHistTosnrInfo(capHistTab)
}
\arguments{
\item{capHistTab}{\itemize{
\item Capture history table of detections
}}
}
\value{
\itemize{
\item SNRInfo data.frame containing SNR, detections, RL, NL
}
}
\description{
Convert capture history DATA.FRAME into the 'SNRinfo' format used
by the callDensity package
capHistTosnrInfo
}
