% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{cde}
\alias{cde}
\title{Call Density Estimate - Main function that accepts parameter file and season#' title: "Call Density"
author: "Brian Miller"
date: "2022-10-12"}
\usage{
cde(p, season, snrDetFun = NULL)
}
\arguments{
\item{p}{A data.frame containing parameters for call density estimation.
Usually created by calling function defaultOutputFileNames}

\item{season}{TimeCode specifying month, season, or year for outputs}

\item{snrDetFun}{OPTIONAL linear-model like structure (GLM,GAM,SCAM,etc)
specifying the SNR-detection function to use. If this is not included,
then the SNR-detection function will be derived from the capture history
table.}
}
\value{
data.frame containing call density inputs, results, and CVs
}
\description{
Density of blue whale D-calls (BmD) in long-term moored recording dataset.
Apply methods of Castro, Harris, et al (in prep) to obtain call density.
}
\details{
$D_c = \\frac{N_c*(1-c)}{kTP_a{\\pi}w^2}$
where:
$D_c$ is call density
$N_c$ is number of calls
$c$ is false discovery rate
$k$ is number of sensors (here always 1)
$T$ is the duration of data analysed
$P_a$ is the probability of detection in the study area
${\\pi}w^2$ is the study area (in km\\^2)
}
