% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{deploymentDuration}
\alias{deploymentDuration}
\title{Calculate deployment duration from recording start times and durations}
\usage{
deploymentDuration(duration, startDate, season = "year")
}
\arguments{
\item{duration}{\itemize{
\item list of recording durations (in s)
}}

\item{startDate}{\itemize{
\item list of start dates for each duration
}}

\item{season}{\itemize{
\item Subset the startDate by this season/month/year (default='year')
}}
}
\value{
\itemize{
\item duration of effort (as numeric)
}
}
\description{
Calculate deployment duration from recording start times and durations
}
