% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{falseDiscoveryCV}
\alias{falseDiscoveryCV}
\title{WARNING: This function is not supported, and probably does not do what
whatever you were hoping it might do. Perhaps consider function c_CV instead.
CV of false discovery rate using Cochran approximation}
\usage{
falseDiscoveryCV(hourlyFalsePosFile, season)
}
\arguments{
\item{hourlyFalsePosFile}{Name of csv file contianing hourly estimates of
false positive rate.}

\item{season}{string or number corresponding to the time of year for which
call densities should be estimated (and data subsetted).}
}
\value{
CV of false discovery rate using Cochran approximation
}
\description{
WARNING: This function is not supported, and probably does not do what
whatever you were hoping it might do. Perhaps consider function c_CV instead.
CV of false discovery rate using Cochran approximation
}
\seealso{
c_CV
}
