% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{falseDiscoveryCV}
\alias{falseDiscoveryCV}
\title{CV of false discovery rate using Cochran approximation}
\usage{
falseDiscoveryCV(p, season)
}
\arguments{
\item{p}{data.frame containing all parameters for call density estimate}

\item{season}{string or number corresponding to the time of year for which
call densities should be estimated (and data subsetted).}
}
\value{
CV of false discovery rate using Cochran approximation
}
\description{
CV of false discovery rate using Cochran approximation
}
