% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{fitSNRdetectionFunc}
\alias{fitSNRdetectionFunc}
\title{Fit an SNR-detection function}
\usage{
fitSNRdetectionFunc(SNRinfo, modelType = "gam", numKnots = 3)
}
\arguments{
\item{SNRinfo}{A data.frame containing detection information in columns
Detected and SNR}

\item{modelType}{one of either 'GAM', 'GLM', or 'SCAM'. ModelType determines
which type of model will be used to estimate the detection-SNR function.}

\item{numKnots}{Number of knots in the GAM or SCAM}
}
\description{
SNR detection functions are binomial models of the form detected ~ SNR. These
can be a GAM (default), GLM, or SCAM.
}
