% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{Nc_CV}
\alias{Nc_CV}
\title{Nc_CV: Coefficient of Variation (CV) of the number of detected calls (Nc).}
\usage{
Nc_CV(Nc, pa, c)
}
\arguments{
\item{Nc}{\itemize{
\item Number of calls detected in the dataset (see function
countDetections())
}}

\item{pa}{\itemize{
\item Average probability of detection in the study area (see function
pDetInArea())
}}

\item{c}{\itemize{
\item false discovery rate (proportion of false positive detections
divided by the total number predicted positive)
}}
}
\value{
cv.Nc - coefficient of variation of the number of detected calls
}
\description{
CV.Nc is calculated from the variance of Nc, which depends on the probability
of detection (pa). Here, Nc represents the number of detected calls, and
independence between detections is assumed.
}
\details{
Rationale for this calculation Variance of a binomial process, pa, is: 1)
sigma^2= n * pa * (1-pa)

We require n, the number of trials conducted, but we have Nc, the number of
positive detections. So we correct for false positives, then scale true
positives by pa to get total number of trials
\enumerate{
\item n =  (Nc*(1-c)/pa; # n: the number of trials conducted We substitute the
right hand side back into 1) to get: 3) sigma^2 = (Nc*(1-c)/pa\emph{pa}(1-pa);
dividing by and multiplying by pa is the same as multiplying by one, leaving
\item sigma^2 = Nc * (1-c) * (1-pa)
}
}
