% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{pa_CV}
\alias{pa_CV}
\title{pa_CV: Coefficient of Variation (CV) of the probability of detection (pa).}
\usage{
pa_CV(pa.all.transects, wt = rep(1, dim(pa.all.transects)[1] - 1))
}
\arguments{
\item{pa.all.transects}{\itemize{
\item matrix containing two columns and same number of
rows as number of transects. The first column contains the mean pa and the
second column contains the standard deviation of pa for each transect. The
final column contains the overall mean and sd.
}}
}
\value{
CV.pa - CV of the probability of deteciton in the area
}
\description{
CV.pa is calculated using the standard deviation of the means and the sum of
the standard deviations of the transects. This approach assumes independence
between transects and homogeneity of error. The summation of SDs assumes a
conservative model.
}
\details{
Follows from Eqn. 6.19 and 6.20 from Harris (2012) PhD Thesis Chapter 6
var_total_Pa = SE_Pa^2 + (sum(SD_Pa_t)/n_t)^2  Eqn.6.19
SE_total_Pa = sqrt(var_total_Pa)               Eqn.6.20

where:

var_total_Pa = variance of the overall mean probability of detection
incorporating all sources of variance
SE_total_Pa = standard error of the overall mean probability of detection
incorporating all sources of variance

SE_Pa = standard error of the n_t transect-specific mean probabilities
of detection
SD_Pa_t = standard deviation of each transect-specific mean probability of
detection
t = transect (1 - n_t)
n_t = total number of transects (8 in Harris's thesis, but adjustable here)

Note: to allow the lengths of transects to vary requires additional weighting
beyond that described by Harris (2012). We account for this with the wt
parameter, which should contain the area for each transect. We adapt Eqn.6.19
So that SE_Pa is derived from the use the weighted mean, and the right-most
term (mean of standard deviations) is also a weighted mean.
}
