% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensityEstimate.R
\name{pa_CV}
\alias{pa_CV}
\title{pa_CV: Coefficient of Variation (CV) of the probability of detection (pa).}
\usage{
pa_CV(pa.all.transects)
}
\arguments{
\item{pa.all.transects}{\itemize{
\item matrix containing two columns and same number of
rows as number of transects. The first column contains the mean pa and the
second column contains the standard deviation of pa for each transect. The
final column contains the overall mean and sd.
}}
}
\value{
CV.pa - CV of the probability of deteciton in the area
}
\description{
CV.pa is calculated using the standard deviation of the means and the sum of
the standard deviations of the transects. This approach assumes independence
between transects and homogeneity of error. The summation of SDs assumes a
conservative model.
}
