% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilityOfDetectionInArea.R
\name{pDetInArea}
\alias{pDetInArea}
\title{Simulation to predict average probability of detection
The simulation has several steps:
(1) SNR regression and TL data now passed in via function call
(a)Thin original TL data so have a 100m range step, not 20m
   The resolution given in the propagation model output may be too fine a
   resolution, and will increase the code run time.
(2) GAM for the SNR regression passed in via function call -
   initial model does not need to be in the loop}
\usage{
pDetInArea(
  res.1,
  SL,
  TLlookup,
  NL,
  transectFile,
  simResultsFile,
  paFile,
  useGLM = FALSE,
  useSCAM = FALSE,
  numKnots = 3,
  output.resolution.m = 100,
  outerloop = 1000
)
}
\arguments{
\item{res.1}{SNR-Detection function from fitSNRdetectionFunc}

\item{SL}{Distribution of source levels (data.frame with columns mean,sd)}

\item{TLlookup}{Transmission loss lookup table with range in the first column
and TL in dB along an azimuth for columns 2-n}

\item{NL}{Noise level distribution (data.frame with columns mean,sd)}

\item{transectFile}{File name base for writing transect outputs}

\item{simResultsFile}{File name base for writing simulation results}

\item{paFile}{File name for writing probability of detection in the area, pa}

\item{useGLM}{Boolean - set true if res.1 contains a GLM}

\item{useSCAM}{Boolean - set true if res.1 contains a SCAM}

\item{numKnots}{Number of knots in GAM or SCAM}

\item{output.resolution.m}{Spatial resolution of density/transect output in m}

\item{outerloop}{Number of iterations in outerloop (default 1000)}
}
\description{
(3) The outer loop (to be run 1000 times): (a) Generate a possible mean and
   standard deviations for source level using a parametric bootstrap approach (b)
   Generate a possible mean and standard deviations for noise level using a
   parametric bootstrap approach (c) Generate a set of coefficients for the
   detector characterisation curve
}
\details{
(4) For each combination of outer loop parameters: For each column of TL
   lookup table : (a) draw ~10000 source level values from a distribution
   generated in outer loop and apply to all of the virtual calls along the
   transect (b) draw ~10000 noise level values from a distribution generated
   in outer loop and apply to all of the range steps along the transect (c)
   using the TL values, calculate the RL given the SL for each virtual call
   (d) calculate the SNR of each virtual call, given the assigned noise
   levels (e) predict prob(detect) given the SNR for all SNR values (~10000
   per profile)

(5) for all p(det), calculate a weighted average (weighting by
   distance) Weight each prob(detect) by the range (i.e. multiply the p(det) by
   the range) Then divide all weighted p(dets) for a transect by the sum of ALL
   ranges across each transect. Having run the outer loop 1000 times, have 1000
   weighted-average p(det) per transect
(6) Final p(det) calculations
 (a) Calculate a transect-specific mean p(det) for each transect
 (b) Calculate an overall p(det) for the whole simulation
 (c) Save results in a results table
(7) Variance calculations
 (a) Calculate the standard error of transect-specific p(det) values
 (b) Calculate the standard deviation of transect-specific p(det) values
 (c) Save results in a table and print the table for future calculations
}
