% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDetFun.R
\name{showDetFun}
\alias{showDetFun}
\title{Plot SNR detection function with observed SNR distributions}
\usage{
showDetFun(model, ...)
}
\arguments{
\item{model}{Fitted detection model from:
fitSNRdetectionFunc(), fitSNRvglm()}

\item{SNRinfo}{Data.frame used to fit the model. Must contain:
\itemize{
\item SNR
\item Detected (logical or 0/1)
}}

\item{distribution}{One of "density", "histogram", "none"}

\item{mirror.height}{Height scaling for mirrored distributions}

\item{npoints}{Number of SNR grid points for prediction}
}
\value{
ggplot object
}
\description{
Visualises fitted SNR detection functions together with the distribution
of detected vs undetected events. Supports mirrored densities or histograms.
}
