% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySupport.R
\name{showSNRdetectionFunc}
\alias{showSNRdetectionFunc}
\title{Plot SNR-detection function (ggplot2)}
\usage{
showSNRdetectionFunc(res.1)
}
\arguments{
\item{res.1}{An SNR-detection function from fitSNRdetectionFunc}
}
\description{
Plot SNR-detection function (ggplot2)
}
