% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySim.R
\name{simCallLocation}
\alias{simCallLocation}
\title{Simulate animals calls using a uniform distribution time and space}
\usage{
simCallLocation(
  n = 1e+06,
  R = 1e+06,
  minDate = Sys.time(),
  maxDate = startDate + 86400
)
}
\arguments{
\item{n}{- Number of calls that will be created.

  For reference 1 whale call every 10 s would be a maximum of 3.1536M calls
  In a continuous year of recording (i.e. 365*24*60*60/10 = 3.1536e6)}

\item{R}{- Radius of study area.}

\item{k}{- Number of sensors (presently always 1)}

\item{endDatetime}{}
}
\description{
Generate a uniform distribution of locations in time and space. The space
component is in two dimensions (x, y), and the time component is in calendar
time (posix.ct).
}
\details{
simCallLocation creates n calls within radius R of the receiver (hydrophone).
This is useful for testing callDensity package with a known call density
(uniform random in circular study area).
}
\examples{
sim <- simCallLocation(n=1e6, R=1e6, minDate = Sys.time(), maxDate=Sys.time()-(365*24*60*60))
}
