% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDensitySim.R
\name{simsTocaptureHistoryTable}
\alias{simsTocaptureHistoryTable}
\title{Create a capture history table from the two simulated detection tables}
\usage{
simsTocaptureHistoryTable(subsampleDet1, subsampleDet2)
}
\arguments{
\item{subsampleDet1}{- Simulated detection table 1}

\item{subsampleDet2}{- Simulated detection table 2}
}
\value{
capture history table containing a row for each detection that was
  detected by either detector.
}
\description{
Create a capture history table from the two simulated detection tables
}
